Safari-609.1.11.1/LayoutTests/media/W3C/video/networkState/networkState_during_loadstart.html

<!doctype html>
<html>
 <head>
  <title>video.networkState - NETWORK_LOADING</title>
  <script src="../../w3cwrapper.js"></script>
 </head>
 <body>
  <p><a href="http://dev.w3.org/html5/spec/Overview.html#dom-media-networkstate">spec reference</a></p>
  <video id="v" autoplay controls>
  </video>
  <div id="log"></div>
  <script>
var t = async_test("videoElement.networkState should be NETWORK_LOADING during loadstart event", {timeout:30000});
var v = document.getElementById("v");
v.addEventListener("loadstart", function() {
  t.step(function() {
   assert_equals(v.networkState,
   v.NETWORK_LOADING);
  });
  t.done();
  v.pause();
});
v.src = getVideoURI("http://media.w3.org/2010/05/video/movie_300") + "?" + new Date() + Math.random();
  </script>
 </body>
</html>
